Move vllm.cmd to vllm.entrypoints.cli
