+++ vllm/scripts.py
@@ -1,8 +1,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
-from vllm.cmd.main import main as vllm_main
+from vllm.entrypoints.cli.main import main as vllm_main
 
 
-# Backwards compatibility for the move from vllm.scripts to vllm.cmd.main
+# Backwards compatibility for the move from vllm.scripts to
+# vllm.entrypoints.cli.main
 def main():
     vllm_main()
